refactor: split SmartSlippageUpdater
